backend/requirements.txt
fastapi
uvicorn
pandas
autogluon.tabular
python-multipart
